Update .bashrc
Removed the zsh section; it was pretty pointless in this file.

Using -vo x11 with mpv alias because mplayer without it frequently crashes for me.

Removed the SUDO_EDITOR check for nano; I never use it and figure most using GitHub are probably not going to use Nano. lol VIM all the way!

Removed vendor from the -o flag with lsblk for the lsblkid alias; it took up too much room and I find it rarely useful.

The chvt part (348-357) now now checks for being on a tty first, since it's surely quicker.