Make git prompt verbose again, but better so
Now the git prompt will more or less keep the same formatting, but
instead of simply saying which branch you're on, the sentence will
include some small information, typically an integer, about the
branch's current state.

There are some areas which need improving, due to inefficiency, but as
it stands, it works well and isn't overly slow.

The next commit will see these changes take effect over at `gitsa()`.

Instead of just saying a branch is clean, I've brought back commit
counts, so now it'll say, for example:

  `Branch 'dev' is 936 commit(s) clean.`

Which I think is aesthetically appealing and interesting to know.

There needs to be a counter for the number of changes to commit in the
message as follows:

  `Branch 'dev' has changes to commit.`

Unfortunately, this could prove quite tricky, but there must be a way.

No longer using the `GIC` variable and the `P` variable concatenation
has been drastically reduced. Although the lines might not look as
pretty (for now), I find it a lot easier to manage.

As result of this new approach, the branch detection block was moved
above the ordered while loops.

A caveat of this commit, is that if the user were to disable the branch
from being displayed, he or she would in-fact break the grey output.
This will soon be addressed, and likely just end up, for example:

  `Changes to commit.`

Or perhaps:

  `Detected changes to commit.`