Add toppy() & various Git functions
I'm happy with `toppy()`. It's essentially a more concise top(1) in
BASH, except it just gives you a straightforward human-readable size
and it sorts it all within BASH using a sorting algo. I convert what
seems to be RSS to human-readable sizes. It's quite pleasant, albeit
fairly simple. Hopefully it continues to work and that I wrote it
properly.