Remove `_Human()` & `inout()`
The human-readable sizes are just not accurate enough with BASH's lack
of floating point arithmetic. It'd be fine if we weren't talking Gs.
The function `_Human()` name might have clashed with BASH completion
scripts, as prepending `_` or `__` is a common convention for functions
therein. I'll just write this in Python or something.